add exception handler

+++ todolist-spring/src/main/java/todolistspring/services/TaskService.java
@@ -1,13 +1,16 @@
 package todolistspring.services;
 
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import todolistspring.dto.TaskDTO;
 import todolistspring.entities.Task;
 import todolistspring.repositories.TaskRepository;
+import todolistspring.services.exceptions.ResourceNotFoundException;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @Service
 public class TaskService {
@@ -23,7 +26,8 @@ public class TaskService {
 
     @Transactional(readOnly = true)
     public TaskDTO findById(Long id) {
-        Task task = taskRepository.findById(id).get();
+        Task task = taskRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Resource not found"));
         return new TaskDTO(task);
     }
 
@@ -40,17 +44,26 @@ public class TaskService {
 
     @Transactional
     public TaskDTO updateById(Long id, TaskDTO taskDTO) {
-        Task task = taskRepository.getReferenceById(id);
-        task.setTitle(taskDTO.title());
-        task.setDescription(taskDTO.description());
-        task.setCompleted(taskDTO.isCompleted());
-        task = taskRepository.save(task);
+        try {
+            Task task = taskRepository.getReferenceById(id);
+            task.setTitle(taskDTO.title());
+            task.setDescription(taskDTO.description());
+            task.setCompleted(taskDTO.isCompleted());
+            task = taskRepository.save(task);
 
-        return new TaskDTO(task);
+            return new TaskDTO(task);
+        }
+        catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Resource not found");
+        }
+
     }
 
     @Transactional
     public void deleteById(Long id) {
+        if (!taskRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Resource not found");
+        }
         taskRepository.deleteById(id);
     }
 }
